Visitors e anotações

+++ src/main/kotlin/Teste.kt
@@ -1,4 +1,4 @@
-class Teste(name: String, num: Int, arr: List<String>) {
+class Teste(name: String, num: Double, arr: List<String>) {
     val name = name
     val num = num
     val arr = arr
